Make a couple of other refactors to get some test coverage on this and make some things easier to reason about
